chore: gitignore prod/staging env files and add example templates
