+++ .gitignore
@@ -65,3 +65,10 @@ cypress.env.json
 
 # Claude Code local settings (personal permissions/paths — not for committing)
 .claude/settings.local.json
+
+# Production and staging environment files contain sensitive keys — never commit
+src/environments/environment.prod.ts
+src/environments/environment.staging.ts
+
+# Auth0 dev environment — requires a real Auth0 tenant, copy from example
+src/environments/environment.auth0-dev.ts
