--- src/environments/environment.prod.ts
@@ -1,31 +1,0 @@
-import packageInfo from '../../package.json';
-export const environment = {
-  production: true,
-  printLogApiUrl: 'https://api.3dprintlog.com',
-  version: packageInfo.version,
-  authentication: {
-    domain: '3dprintlog.auth0.com',
-    client_id: '3JOtliMnZIQJDZ5auHia2GZEydfXW9xB',
-    audience: 'https://3dprintlog.com/api',
-  },
-  googleAnalyticsId: 'G-4TEMNSY6QX',
-  appInsights: {
-    instrumentationKey: '3c9f6914-555b-4442-8724-0d5b706d115a',
-  },
-  /**
-   * Feature Flags
-   */
-  features: {
-    /**
-     * Enable the UserProfile Feature
-     */
-    userProfile: true,
-  },
-  googleAds: {
-    trafficSearchConversion: 'AW-837103521/Zn4kCLWd85IDEKHflI8D',
-  },
-  stripe: {
-    proMonthlyPriceId: 'price_1T5tO0JyHW0UWXhycxb6dpjw',
-    proAnnualPriceId: 'price_1T5tO1JyHW0UWXhyKi6DsqpF',
-  },
-};
